Fix bugs where oai_hss would not build on centos 8.3 due to missing libuv.
On centos 8.3, the configure script for cpp_driver for would not
properly detect the version of libuv which comes with centos 8.3 and
would fail to build.

To resolve this, we modify build_helper.hss_rel14 to call build_libuv()
to build libuv from source, however build_libuv() also has to be
modified to use autoconf/configure/make to build libuv instead of
cmake.  The reason for this is that the cmake method does not install
a versioned version of the libuv library, so when the oai_hss binary
is run it would complain that it could not find libuv.so (note without
the version number...).